Gjeninnfører separate komponenter for mobil/desktop hovedmeny
