+++ src/komponenter/header/header-regular/HeaderMenylinje.less
@@ -19,13 +19,6 @@
         height: 100%;
     }
 
-    &__menyer {
-        display: flex;
-        flex-wrap: wrap;
-        width: 100%;
-        height: 0;
-    }
-
     &__spacer {
         flex-grow: 1;
         flex-shrink: 0;
