+++ src/komponenter/header/header-regular/common/knapper/logg-inn-knapp/LoggInnKnapp.tsx
@@ -8,21 +8,18 @@ import { MenuValue } from 'utils/meny-storage-utils';
 import KnappBase from 'nav-frontend-knapper';
 import './LoggInnKnapp.less';
 
-type Props = {
-    type?: 'standard' | 'hoved' | 'fare' | 'flat';
-};
-
 export const loginKnappId = 'desktop-login-knapp';
 
-export const LoggInnKnapp = ({ type }: Props) => {
-    const { environment } = useSelector((state: AppState) => state);
-    const { PARAMS } = useSelector((state: AppState) => state.environment);
-    const language = useSelector((state: AppState) => state.language.language);
-    const { authenticated } = useSelector(
-        (state: AppState) => state.innloggingsstatus.data
-    );
-    const arbeidsflate = useSelector(
-        (state: AppState) => state.arbeidsflate.status
+const stateSelector = (state: AppState) => ({
+    innlogget: state.innloggingsstatus.data.authenticated,
+    arbeidsflate: state.arbeidsflate.status,
+    language: state.language.language,
+    environment: state.environment,
+});
+
+export const LoggInnKnapp = () => {
+    const { innlogget, arbeidsflate, language, environment } = useSelector(
+        stateSelector
     );
 
     const handleButtonClick = () => {
@@ -30,36 +27,35 @@ export const LoggInnKnapp = ({ type }: Props) => {
         const { MINSIDE_ARBEIDSGIVER_URL } = environment;
         const appUrl = location.origin + location.pathname;
         const loginUrl = `${
-            PARAMS.REDIRECT_TO_APP || erNavDekoratoren()
+            environment.PARAMS.REDIRECT_TO_APP || erNavDekoratoren()
                 ? `${LOGIN_URL}/login?redirect=${appUrl}`
                 : arbeidsflate === MenuValue.ARBEIDSGIVER
                 ? `${LOGIN_URL}/login?redirect=${MINSIDE_ARBEIDSGIVER_URL}`
                 : `${LOGIN_URL}/login?redirect=${DITT_NAV_URL}`
-        }&level=${PARAMS.LEVEL}`;
+        }&level=${environment.PARAMS.LEVEL}`;
 
         gaEvent({
             context: arbeidsflate,
             category: GACategory.Header,
-            action: authenticated ? 'logg-ut' : 'logg-inn',
+            action: innlogget ? 'logg-ut' : 'logg-inn',
         });
 
-        return authenticated
+        return innlogget
             ? (window.location.href = LOGOUT_URL)
             : (window.location.href = loginUrl);
     };
 
     const knappetekst = finnTekst(
-        authenticated ? 'logg-ut-knapp' : 'logg-inn-knapp',
+        innlogget ? 'logg-ut-knapp' : 'logg-inn-knapp',
         language
     );
 
     return (
         <div className={'login-knapp-container'}>
             <KnappBase
-                className={`login-knapp${authenticated ? ' logout-knapp' : ''}`}
+                className={`login-knapp${innlogget ? ' logout-knapp' : ''}`}
                 onClick={handleButtonClick}
                 id={loginKnappId}
-                type={type}
             >
                 {knappetekst}
             </KnappBase>
