+++ src/komponenter/header/header-regular/desktop/DesktopHeader.less
@@ -11,7 +11,8 @@
         max-width: @desktopMaxWidth;
     }
 
-    .mobilmeny {
+    .mobilmeny,
+    #mobilmeny-knapp-id {
         display: none;
     }
