+++ src/komponenter/header/header-regular/desktop/hovedmeny/HovedmenyDesktop.tsx
@@ -1,80 +1,67 @@
 import React from 'react';
-import { useState } from 'react';
-import { useEffect } from 'react';
+import { useSelector } from 'react-redux';
+import { Status } from 'api/api';
+import { AppState } from 'store/reducers';
+import { getHovedmenyNode } from 'utils/meny-storage-utils';
+import EkspanderbarMeny from 'komponenter/header/header-regular/common/ekspanderbar-meny/EkspanderbarMeny';
+import Spinner from 'komponenter/header/header-regular/common/spinner/Spinner';
 import { KbNavMain } from 'utils/keyboard-navigation/useKbNavMain';
-import { KbNavConfig } from 'utils/keyboard-navigation/kb-navigation-setup';
-import { configForNodeGroup } from 'utils/keyboard-navigation/kb-navigation-setup';
-import { useKbNavSub } from 'utils/keyboard-navigation/useKbNavSub';
-import { KbNavGroup } from 'utils/keyboard-navigation/kb-navigation';
-import { matchMedia } from 'utils/match-media-polyfill';
-import { getHovedmenyMaxColsPerRow } from 'utils/keyboard-navigation/kb-navigation-setup';
-import { MenuValue } from 'utils/meny-storage-utils';
-import { Language } from 'store/reducers/language-duck';
-import { MenyNode } from 'store/reducers/menu-duck';
-import { Toppseksjon } from 'komponenter/header/header-regular/desktop/hovedmeny/topp-seksjon/Toppseksjon';
-import { Hovedseksjon } from 'komponenter/header/header-regular/desktop/hovedmeny/hoved-seksjon/Hovedseksjon';
-import { Bunnseksjon } from 'komponenter/header/header-regular/desktop/hovedmeny/bunn-seksjon/Bunnseksjon';
+import HovedmenyDesktopInnhold from 'komponenter/header/header-regular/desktop/hovedmeny/HovedmenyDesktopInnhold';
+import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/knapper/hovedmeny-knapp/HovedmenyKnapp';
 
 const classname = 'desktop-hovedmeny';
+export const desktopHovedmenyKnappId = 'desktop-hovedmeny-knapp-id';
 
-const nodeGroup = KbNavGroup.Hovedmeny;
-const mqlScreenWidth = matchMedia('(min-width: 1024px)');
+const stateSelector = (state: AppState) => ({
+    arbeidsflate: state.arbeidsflate.status,
+    menyPunkter: state.menypunkt,
+    language: state.language.language,
+    isOpen: state.dropdownToggles.hovedmeny,
+});
 
 type Props = {
-    arbeidsflate: MenuValue;
-    menyPunkter?: MenyNode;
-    language: Language;
-    isOpen: boolean;
     kbNavMainState: KbNavMain;
 };
 
-export const HovedmenyDesktop = ({
-    kbNavMainState,
-    arbeidsflate,
-    menyPunkter,
-    language,
-    isOpen,
-}: Props) => {
-    const kbConfig = configForNodeGroup[nodeGroup];
-    const [kbNavConfig, setKbNavConfig] = useState<KbNavConfig>(kbConfig);
-    useKbNavSub(kbNavConfig, kbNavMainState, isOpen);
+export const HovedmenyDesktop = ({ kbNavMainState }: Props) => {
+    const { arbeidsflate, menyPunkter, language, isOpen } = useSelector(
+        stateSelector
+    );
 
-    const updateMaxCols = () =>
-        setKbNavConfig({
-            ...kbNavConfig,
-            maxColsPerRow: getHovedmenyMaxColsPerRow(classname),
-        });
+    const hovedmenyPunkter = getHovedmenyNode(
+        menyPunkter.data,
+        language,
+        arbeidsflate
+    );
 
-    useEffect(() => {
-        mqlScreenWidth.addEventListener('change', updateMaxCols);
-        return () => {
-            mqlScreenWidth.removeEventListener('change', updateMaxCols);
-        };
-    }, []);
-
-    useEffect(() => {
-        updateMaxCols();
-    }, [menyPunkter, arbeidsflate]);
-
-    if (!menyPunkter) {
+    // Hide empty menues
+    if (menyPunkter.status === Status.OK && !hovedmenyPunkter?.hasChildren) {
         return null;
     }
 
     return (
-        <div className={classname}>
-            <Toppseksjon classname={classname} />
-            <Hovedseksjon
-                menyLenker={menyPunkter}
-                classname={classname}
+        <>
+            <HovedmenyKnapp id={desktopHovedmenyKnappId} />
+            <EkspanderbarMeny
                 isOpen={isOpen}
-            />
-            <Bunnseksjon
                 classname={classname}
-                language={language}
-                arbeidsflate={arbeidsflate}
-            />
-        </div>
+                id={classname}
+            >
+                {menyPunkter.status === Status.OK ? (
+                    <HovedmenyDesktopInnhold
+                        arbeidsflate={arbeidsflate}
+                        isOpen={isOpen}
+                        language={language}
+                        menyPunkter={hovedmenyPunkter}
+                        kbNavMainState={kbNavMainState}
+                    />
+                ) : (
+                    <Spinner
+                        tekstId={'meny-loading'}
+                        className={isOpen ? 'spinner-container--active' : ''}
+                    />
+                )}
+            </EkspanderbarMeny>
+        </>
     );
 };
-
-export default HovedmenyDesktop;
