+++ src/komponenter/header/header-regular/mobil/MobilHeader.less
@@ -8,7 +8,8 @@
     .desktop-sok-dropdown,
     .desktop-minside-meny__knapp,
     .desktop-minside-meny,
-    .logout-knapp {
+    .logout-knapp,
+    #desktop-hovedmeny-knapp-id {
         display: none;
     }
 
@@ -23,8 +24,7 @@
         margin-right: 0.5rem;
     }
 
-    .hovedmeny__knapp {
+    #mobilmeny-knapp-id {
         margin-right: @margins-sides-mobil;
-        order: 1;
     }
 }
