+++ src/komponenter/header/header-regular/mobil/hovedmeny/HovedmenyMobil.tsx
@@ -7,6 +7,13 @@ import { GACategory, gaEvent } from 'utils/google-analytics';
 import { toggleUndermenyVisning } from 'store/reducers/dropdown-toggle-duck';
 import { toggleHovedmeny } from 'store/reducers/dropdown-toggle-duck';
 import { dataInitState } from 'store/reducers/menu-duck';
+import { HovedmenyKnapp } from 'komponenter/header/header-regular/common/knapper/hovedmeny-knapp/HovedmenyKnapp';
+import EkspanderbarMeny from 'komponenter/header/header-regular/common/ekspanderbar-meny/EkspanderbarMeny';
+import { Status } from 'api/api';
+import Spinner from 'komponenter/header/header-regular/common/spinner/Spinner';
+
+export const mobilmenyKnappId = 'mobilmeny-knapp-id';
+const classname = 'mobilmeny';
 
 const stateSelector = (state: AppState) => ({
     meny: state.menypunkt,
@@ -54,18 +61,35 @@ export const HovedmenyMobil = () => {
     const isOpen = hovedIsOpen || underIsOpen || varselIsOpen;
 
     return (
-        <MobilVisningsmeny
-            classname={'mobilmeny'}
-            menyLenker={hovedmenyPunkter || dataInitState}
-            minsideLenker={
-                getMinsideMenyNode(meny.data, language) || dataInitState
-            }
-            menuIsOpen={hovedIsOpen}
-            underMenuIsOpen={underIsOpen}
-            varslerIsOpen={varselIsOpen}
-            togglemenu={menutoggle}
-            togglehovedmenu={hovedmenutoggle}
-            lang={language}
-        />
+        <>
+            <HovedmenyKnapp id={mobilmenyKnappId} />
+            <EkspanderbarMeny
+                isOpen={hovedIsOpen}
+                classname={classname}
+                id={classname}
+            >
+                {meny.status === Status.OK ? (
+                    <MobilVisningsmeny
+                        classname={classname}
+                        menyLenker={hovedmenyPunkter || dataInitState}
+                        minsideLenker={
+                            getMinsideMenyNode(meny.data, language) ||
+                            dataInitState
+                        }
+                        menuIsOpen={hovedIsOpen}
+                        underMenuIsOpen={underIsOpen}
+                        varslerIsOpen={varselIsOpen}
+                        togglemenu={menutoggle}
+                        togglehovedmenu={hovedmenutoggle}
+                        lang={language}
+                    />
+                ) : (
+                    <Spinner
+                        tekstId={'meny-loading'}
+                        className={isOpen ? 'spinner-container--active' : ''}
+                    />
+                )}
+            </EkspanderbarMeny>
+        </>
     );
 };
